Code formatting, refactoring and fixes.

+++ benchmarks/crand.c
@@ -4,7 +4,7 @@
 int main(void)
 {
     printf("RAND_MAX = %d\n", RAND_MAX);
-    for (unsigned long long i = 0; i < 1000000000; ++i)
+    for (uint64_t i = 0; i < 1000000000; ++i)
     {
         rand();
     }
